requirements.txt
amqp==5.3.1
aniso8601==9.0.1
axios==0.4.0
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.2
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
npm==0.1.1
optional-django==0.1.0
passlib==1.7.4
prompt_toolkit==3.0.48
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.0
requests==2.32.3
rich==13.9.4
setuptools==75.3.0
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.2
WTForms==3.2.1
